+++ CHANGELOG.md
@@ -8,11 +8,13 @@
     - moved Measurements -> MeasurementsFeature (extends EnlightenFeature)
     - fixed EtalonCorrection
     - shortcuts
+        - ctrl-M toggles library Matching
         - ctrl-Q quits
         - ctrl-shift-D toggles DALAI
-        - ctrl-shift-L toggles Library Matching
         - ctrl-shift-S exports all Saved measurements
         - ctrl-up/dn scrolls control palette or hardware view as appropriate
+    - plugins
+        - added Prod.ConfigurationCheck
 - 2026-07-08 4.2.5
     - fixed poorly-tested bugs in 4.2.4
         - including inability to write horiz ROI in EEPROM